Fix "dial: Ssl::PeerConnector::sslCrtvdHandleReply threw exception: callback != NULL"
This bug investigated after the r14528 patch. 
The (forgotten?) Ssl::PeekingPeerConnector::callback member is always NULL and 
hides the callback member of the parent Ssl::PeerConnector class. This is has
as result the failure of "Must(callback != NULL)" clause inside the 
Ssl::PeekingPeerConnector::tunnelInsteadOfNegotiating method investigated with 
the r14528 patch.

This is a Measurement Factory project